Moved sync-creation tests to the (new) pegl.sync test module
